Add support for rubies compiled with `-DUSE_FLONUM=0`
